Add github actions, update test runner, upgrade linting, add npm publishing
