+++ .github/workflows/ci.yml
@@ -1,0 +1,20 @@
+name: CI
+on: pull_request
+jobs:
+  test:
+    timeout-minutes: 5
+    runs-on: ubuntu-latest
+    steps:
+      - uses: Brightspace/third-party-actions@actions/checkout
+      - uses: Brightspace/third-party-actions@actions/setup-node
+        with:
+          node-version-file: .nvmrc
+      - name: Install dependencies
+        run: |
+          npm install
+          npm install @web/test-runner-playwright --no-save
+          npx playwright install-deps
+      - name: Lint
+        run: npm run lint
+      - name: Unit Tests
+        run: npx web-test-runner test/**/*.test.js --node-resolve --group default --playwright --browsers chromium firefox webkit
