--- .travis.yml
@@ -1,19 +1,0 @@
-language: node_js
-node_js: node
-addons:
-  chrome: stable
-script:
-- npm run test:lint
-- |
-  if [ $TRAVIS_PULL_REQUEST != false ] && [ $TRAVIS_SECURE_ENV_VARS == true ]; then
-    echo "Pull request with secure environment variables, running Sauce tests...";
-    npm run test:polymer:sauce || travis_terminate 1;
-  else
-    echo "Not a pull request and/or no secure environment variables, running headless tests...";
-    npm run test:polymer:local || travis_terminate 1;
-  fi
-env:
-  global:
-  - SAUCE_USERNAME: Desire2Learn
-  # SAUCE_ACCESS_KEY
-  - secure: RssTw9EWnkVNg5oeg48gvyRgizU/BbEsO+yEL5DKzbIhHZUqKzdsVJQMuyLr2PYGiV/1NSDAQyBE0WkaZG46IZJv1dfYmQi+OxPL9cMveLAQhZWTvSQM3pwsXbDuHlZ0w45+Q7+jKksfTEzow1pafcqLgGntlttVeJaujyr8Pko4dmDHZdrEFudU8gIwhL/sQ9cHzAwe46m8aXpLM083efI/hAUKAbjFfge0f5s8FpABErZ6xz7YAmrMM4dDm0laGN7TShBMLAUipilzn4XC3YzJfjNdoGMat/Yqb0ghXA/8JjRp+vWCq2yww8s3f+8MXJUBBejyQvyrj/30IISdv9S02FF/IEdmBbxvVqJIJm+XvQ4aM2i5e78zGpFEwNNAzEYLK6cRgb79HIJHjsK6Zcy5TW2UOuWxQs4MKp1I4R3sYdwHTETwjNEQe4LQk3Jef+RcUH/eOJHTyMhBUA0GgUIX45Pq3A5vDjtT4mfbqO9f0sM3CMpsm1WMnT/xcYQcSFoHLgeTBCDFG86wYgFuu2bHOXsiPqMHr08upGpUCa92l4hCuVnPvskYRw8W8bvHf/c4+C4pKyLE78HyL8ElQzQ738wB+cfAhFkO7XzwZkJ3Sp0jEK1hkYWeneuRyrYo3fQdyUmWezzf2JsAnuE+YgtJMUPfzIRrIiLQi8m29Q4=
