+++ package.json
@@ -1,43 +1,41 @@
 {
   "description": "Shared Polymer behaviors for use in web components",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/Brightspace/polymer-siren-behaviors.git"
   },
   "name": "d2l-polymer-siren-behaviors",
   "scripts": {
-    "test": "npm run test:lint && npm run test:polymer:local",
-    "test:lint": "npm run test:lint:wc && npm run test:lint:js",
-    "test:lint:js": "eslint . --ext .js,.html test/**/*.js test/**/*.html demo/**/*.js demo/**/*.html",
-    "test:lint:wc": "polymer lint",
-    "test:polymer:local": "polymer test --skip-plugin sauce",
-    "test:polymer:sauce": "polymer test --skip-plugin local"
+    "test": "npm run test:lint && npm run test:headless",
+    "test:headless": "web-test-runner test/**/*.test.js --node-resolve --group default",
+    "test:headless:watch": "web-test-runner test/**/*.test.js --node-resolve --group default --watch",
+    "test:lint": "npm run test:lint:js",
+    "test:lint:js": "eslint . --ext .js,.html test/**/*.js"
   },
   "author": "D2L Corporation",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@babel/polyfill": "^7.0.0",
-    "@polymer/iron-component-page": "^3.0.0-pre.18",
-    "@polymer/iron-demo-helpers": "^3.0.0-pre.18",
-    "@polymer/polymer": "^3.0.0",
-    "@webcomponents/webcomponentsjs": "^2.2.1",
-    "babel-eslint": "^10.0.1",
-    "eslint": "^4.15.0",
-    "eslint-config-brightspace": "^0.4.0",
-    "eslint-plugin-html": "^4.0.1",
-    "lie": "^3.3.0",
-    "polymer-cli": "^1.9.1",
-    "wct-browser-legacy": "^1.0.1",
-    "whatwg-fetch": "^2.0.0"
+    "@babel/core": "^7",
+    "@babel/eslint-parser": "^7",
+    "@open-wc/testing": "^3",
+    "@web/test-runner": "^0.13",
+    "eslint": "^8",
+    "eslint-config-brightspace": "^0.16",
+    "eslint-plugin-html": "^6",
+    "eslint-plugin-import": "^2",
+    "eslint-plugin-lit": "^1",
+    "eslint-plugin-sort-class-members": "^1",
+    "sinon": "^14"
   },
   "version": "1.4.0",
-  "resolutions": {
-    "inherits": "2.0.3",
-    "samsam": "1.1.3",
-    "supports-color": "3.1.2",
-    "type-detect": "1.0.0"
-  },
+  "files": [
+    "/store",
+    "entity-loading-behavior.js",
+    "siren-entity-loading.js"
+  ],
   "dependencies": {
+    "@polymer/polymer": "^3.0.0",
     "d2l-fetch": "Brightspace/d2l-fetch.git#semver:^2",
     "fastdom": "^1.0.8",
     "siren-parser": "^8.0.0"
