--- polymer.json
@@ -1,8 +1,0 @@
-{
-  "npm": true,
-  "lint": {
-    "rules": [
-      "polymer-3"
-    ]
-  }
-}
