+++ test/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+    "extends": "brightspace/open-wc-testing-config"
+}
