--- test/entity-behavior.html
@@ -1,29 +1,0 @@
-<!doctype html>
-
-<html>
-	<head>
-		<title>d2l-rubric test</title>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1, user-scalable=yes">
-		<script src="../../@babel/polyfill/browser.js"></script>
-		<script src="../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-
-		<!-- For IE11 -->
-		<script src="../../lie/dist/lie.polyfill.min.js"></script>
-		<!-- For IE11 -->
-		<script src="../../whatwg-fetch/fetch.js"></script>
-		
-		<script src="../../wct-browser-legacy/browser.js"></script>
-		<script src="./utilities/fetch-siren-entity-whitelist.js"></script>
-
-		<script type="module" src="./entity-behavior-test-component.js"></script>
-	</head>
-	<body>
-		<test-fixture id="basic">
-			<template>
-				<entity-behavior-test-component></entity-behavior-test-component>
-			</template>
-		</test-fixture>
-		<script src="./entity-behavior.js"></script>
-	</body>
-</html>
