--- test/entity-store.html
@@ -1,23 +1,0 @@
-<!doctype html>
-
-<html>
-	<head>
-		<title>d2l-rubric test</title>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1, user-scalable=yes">
-		<script src="../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-
-		<!-- For IE11 -->
-		<script src="../../lie/dist/lie.polyfill.min.js"></script>
-		<!-- For IE11 -->
-		<script src="../../whatwg-fetch/fetch.js"></script>
-
-		<script src="../../wct-browser-legacy/browser.js"></script>
-		<script src="./utilities/fetch-siren-entity-whitelist.js"></script>
-		<script type="module" src="../store/entity-store.js"></script>
-
-	</head>
-	<body>
-		<script src="./entity-store.js"></script>
-	</body>
-</html>
