--- test/index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
-	<script src="../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-
-	<!-- For IE11 -->
-	<script src="../../lie/dist/lie.polyfill.min.js"></script>
-	<!-- For IE11 -->
-	<script src="../../whatwg-fetch/fetch.js"></script>
-
-	<script src="../../wct-browser-legacy/browser.js"></script>
-</head>
-
-<body>
-	<script>
-		/* global WCT */
-		'use strict';
-		WCT.loadSuites([
-			'./entity-behavior.html',
-			'./entity-store.html',
-			'./siren-action-behavior.html',
-			'./action-queue.html',
-			'./siren-entity-loading.html'
-		]);
-	</script>
-</body>
-
-</html>
