--- test/siren-entity-loading.html
@@ -1,30 +1,0 @@
-<!doctype html>
-
-<html>
-	<head>
-		<title>siren-entity-loading test</title>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1, user-scalable=yes">
-		<script src="../../@babel/polyfill/browser.js"></script>
-		<script src="../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-
-		<!-- For IE11 -->
-		<script src="../../lie/dist/lie.polyfill.min.js"></script>
-		<!-- For IE11 -->
-		<script src="../../whatwg-fetch/fetch.js"></script>
-		
-		<script src="../../wct-browser-legacy/browser.js"></script>
-
-		<script type="module" src="../siren-entity-loading.js"></script>
-	</head>
-	<body>
-		<test-fixture id="basic">
-			<template>
-				<siren-entity-loading></siren-entity-loading>
-			</template>
-		</test-fixture>
-		<script src="./siren-entity-loading.js">
-
-		</script>
-	</body>
-</html>
