+++ test/static-data/simple-collection/items/0.json
@@ -5,9 +5,9 @@
     },
     "links": [{
         "rel": ["collection"],
-        "href": "static-data/simple-collection/collection.json"
+        "href": "test/static-data/simple-collection/collection.json"
     }, {
         "rel": ["self"],
-        "href": "static-data/simple-collection/items/0.json"
+        "href": "test/static-data/simple-collection/items/0.json"
     }]
 }
