--- test/utilities/fetch-siren-entity-whitelist.js
@@ -1,9 +1,0 @@
-/* global sinon D2L */
-'use strict';
-
-function stubWhitelist() { // eslint-disable-line no-unused-vars
-	var isWhitelisted = sinon.stub();
-	isWhitelisted.returns(true);
-
-	(D2L.PolymerBehaviors.FetchSirenEntityBehavior || {})._isWhitelisted = isWhitelisted;
-}
