--- wct.conf.json
@@ -1,49 +1,0 @@
-{
-	"plugins": {
-		"local": {
-			"browsers": ["chrome"],
-			"browserOptions": {
-				"chrome": ["headless", "disable-gpu", "no-sandbox"]
-			}
-		},
-		"sauce": {
-			"browsers": [
-				{
-					"browserName": "chrome",
-					"platform": "OS X 10.13",
-					"version": ""
-				},
-				{
-					"browserName": "chrome",
-					"platform": "Windows 10",
-					"version": ""
-				},
-				{
-					"browserName": "firefox",
-					"platform": "OS X 10.13",
-					"version": ""
-				},
-				{
-					"browserName": "firefox",
-					"platform": "Windows 10",
-					"version": ""
-				},
-				{
-					"browserName": "safari",
-					"platform": "OS X 10.13",
-					"version": ""
-				},
-				{
-					"browserName": "microsoftedge",
-					"platform": "Windows 10",
-					"version": ""
-				},
-				{
-					"browserName": "internet explorer",
-					"platform": "Windows 10",
-					"version": "11"
-				}
-			]
-		}
-	}
-}
